backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.9.0
httpx==0.27.2
python-multipart==0.0.12
aiofiles==24.1.0
